+++ packages/render-ui/src/color.js
@@ -49,6 +49,9 @@ export const defaults = {
   KEYPAD_BUTTON: 'rgb(188, 194, 229)',
   KEYPAD_BUTTON_OPERATOR: 'rgb(255, 159, 192)',
   KEYPAD_EMPTY_PLACEHOLDER: 'rgba(245, 0, 87, 0.4)',
+  // these are used for graphing UI elements
+  BUTTON_BORDER: 'rgba(0, 0, 0, 0.23)',
+  BUTTON_HOVER_BG: 'rgba(0, 0, 0, 0.08)',
 };
 
 Object.freeze(defaults);
@@ -118,10 +121,13 @@ export const blueGrey900 = () => pv('blue-grey-900', defaults.BLUE_GREY900);
 export const keypadButton = () => pv('keypad-button', defaults.KEYPAD_BUTTON);
 export const keypadButtonOperator = () => pv('keypad-button-operator', defaults.KEYPAD_BUTTON_OPERATOR);
 export const keypadEmptyPlaceholder = () => pv('keypad-empty-placeholder', defaults.KEYPAD_EMPTY_PLACEHOLDER);
+export const buttonBorder = () => pv('button-border', defaults.BUTTON_BORDER);
+export const buttonHoverBg = () => pv('button-hover-bg', defaults.BUTTON_HOVER_BG);
 
 export const visualElementsColors = {
   AXIS_LINE_COLOR: '#5A53C9',
   ROLLOVER_FILL_BAR_COLOR: '#050F2D',
   GRIDLINES_COLOR: '#8E88EA',
   PLOT_FILL_COLOR: '#1463B3',
+  SHAPES_FILL_COLOR: '#7986cb',
 };