+++ packages/plot/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.2.0-next.18](https://github.com/pie-framework/pie-lib/compare/@pie-lib/plot@3.2.0-next.17...@pie-lib/plot@3.2.0-next.18) (2026-03-18)
+
+**Note:** Version bump only for package @pie-lib/plot
+
+# [3.2.0-next.17](https://github.com/pie-framework/pie-lib/compare/@pie-lib/plot@3.2.0-next.16...@pie-lib/plot@3.2.0-next.17) (2026-03-18)
+
+**Note:** Version bump only for package @pie-lib/plot
+
 # [3.2.0-next.16](https://github.com/pie-framework/pie-lib/compare/@pie-lib/plot@3.2.0-next.15...@pie-lib/plot@3.2.0-next.16) (2026-03-11)
 
 **Note:** Version bump only for package @pie-lib/plot